Widen GPU synapse from_index to u32 after neat-core u16 narrowing
neat-core narrowed `SynapseData::from_index` from u32 to u16 (core
Issue #177). The batched GPU path in `build_batched_network_data` still
assigned it straight into `SynapseGpu.from_index` (u32), so
`rust_scorer` failed to compile with E0308 (expected u32, found u16),
breaking the "Unit tests + coverage" job on every downstream PR.

Widen losslessly with `u32::from(s.from_index)`, matching the existing
`squash_type`/`num_synapses` conversions. The GPU SSBO and WGSL
`SynapseGpu` struct keep their u32 layout (WGSL has no u16).

Unblocks stSoftwareAU/NEAT-AI-Examples#604 CI.

Vibe-Coder-Run-Id: vibe-mqqdqjwm-560862

+++ rust_scorer/src/gpu/forward_mse_batched.rs
@@ -225,7 +225,10 @@ pub fn build_batched_network_data(
         for s in &net.synapses {
             synapses.push(SynapseGpu {
                 weight: s.weight,
-                from_index: s.from_index,
+                // neat-core narrowed `SynapseData::from_index` to `u16` (core
+                // Issue #177); widen losslessly to the `u32` the GPU SSBO and
+                // WGSL `SynapseGpu` expect, matching the conversions above.
+                from_index: u32::from(s.from_index),
             });
         }
